+++ testing/fetcher.ts
@@ -229,15 +229,16 @@ export class AuthenticatingFetcher implements Fetcher {
     const {clientId, clientSecret, accessToken, refreshToken, scopes} = this._credentials as OAuth2Credentials;
     assertCondition(accessToken);
     assertCondition(refreshToken);
-    const {tokenUrl} = this._authDef;
+    const {tokenUrl, resource} = this._authDef;
     if (!tokenUrl) {
       throw new Error('Dynamic Client Registration (DCR) is not supported when testing locally');
     }
 
-    const params = {
+    const params: {[key: string]: string | undefined} = {
       grant_type: 'refresh_token',
       refresh_token: refreshToken,
       client_id: clientId,
+      resource,
     };
 
     const headers = new Headers({
@@ -248,6 +249,9 @@ export class AuthenticatingFetcher implements Fetcher {
     const formParams = new URLSearchParams();
     const formParamsWithSecret = new URLSearchParams();
     for (const [key, value] of Object.entries(params)) {
+      if (value === undefined) {
+        continue;
+      }
       formParams.append(key, value.toString());
       formParamsWithSecret.append(key, value.toString());
     }